refactor(browser): extract headed scheduler
